+++ ExtractData/Gyro.cs
@@ -20,7 +20,7 @@ namespace ExtractData
             Z = new Data("Z");
         }
 
-        public void addData(int pos, int data)
+        public void addData(int pos, double data)
         {
             switch (pos)
             {
